clean sample 02
