+++ 02-Custom-Login/app/src/main/AndroidManifest.xml
@@ -13,17 +13,13 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
-        <activity android:name=".activities.StartActivity">
+        <activity android:name=".LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-
-        <activity android:name=".activities.LoginActivity" />
-        <activity android:name=".activities.MainActivity"/>
+        <activity android:name=".MainActivity" />
     </application>
 
 </manifest>
