+++ 02-Custom-Login/app/src/main/java/com/auth0/customlogindemo/LoginActivity.java
@@ -1,0 +1,60 @@
+package com.auth0.customlogindemo;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+
+import com.auth0.android.Auth0;
+import com.auth0.android.authentication.AuthenticationAPIClient;
+import com.auth0.android.authentication.AuthenticationException;
+import com.auth0.android.callback.BaseCallback;
+import com.auth0.android.result.Credentials;
+
+
+public class LoginActivity extends Activity {
+
+    private static final String DEFAULT_DB_CONNECTION = "Username-Password-Authentication";
+
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+
+        // Bind your views
+        final EditText emailEditText = (EditText) findViewById(R.id.emailEditext);
+        final EditText passwordEditText = (EditText) findViewById(R.id.passwordEditext);
+        Button loginButton = (Button) findViewById(R.id.loginButton);
+
+        // Add the onClick listener to the login
+        loginButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                //Show a progress dialog to block the UI while the request is being made.
+                login(emailEditText.getText().toString(), passwordEditText.getText().toString());
+            }
+        });
+
+
+    }
+
+    private void login(String email, String password) {
+        Auth0 auth0 = new Auth0(getString(R.string.auth0_client_id), getString(R.string.auth0_domain));
+        AuthenticationAPIClient client = new AuthenticationAPIClient(auth0);
+
+        client.login(email, password, DEFAULT_DB_CONNECTION).start(new BaseCallback<Credentials, AuthenticationException>() {
+            @Override
+            public void onSuccess(Credentials payload) {
+                startActivity(new Intent(LoginActivity.this, MainActivity.class));
+            }
+
+            @Override
+            public void onFailure(AuthenticationException error) {
+                //Show error to the user
+            }
+        });
+    }
+}
+
+
