--- 02-Custom-Login/app/src/main/java/com/auth0/customlogindemo/activities/StartActivity.java
@@ -1,23 +1,0 @@
-package com.auth0.customlogindemo.activities;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import com.auth0.customlogindemo.R;
-
-
-public class StartActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_start);
-
-        Intent lockIntent = new Intent(this, LoginActivity.class);
-        startActivity(lockIntent);
-
-    }
-
-
-}
